pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "posts-api"
version = "0.1.0"
authors = [
    { name = "Test", email = "test@example.com" },
]
description = "Posts API with async SQLAlchemy"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "asyncpg",
    "python-dotenv",
    "pytest",
    "pytest-asyncio",
    "httpx",
]
